% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{APE}
\alias{APE}
\title{Multi-fold cross validation for the MXFAR model}
\usage{
APE(SeriesNum, Tlength, y, u, p, d, r, Q, bwp = 0.1, numpoints = 50)
}
\arguments{
\item{SeriesNum}{A numeric vector stating the number of series in a 'group' data}

\item{Tlength}{A numeric scalar stating the number of time points per series}

\item{y}{A matrix containing the 'stacked' multivariate time series data from multiple series}

\item{u}{A vector containing the 'stacked' reference signal series}

\item{p}{The order of MXFAR(p) model}

\item{d}{The lag of the reference signal}

\item{r}{The number of time points for predictions}

\item{Q}{The number of segments considered for the time series}

\item{bwp}{The bandwidth proportion from 0 to 1 (with respect to the range of observations)}

\item{numpoints}{The number of discretized points in the range of the reference signal 'u' that will be used to evaluate the local linear approximation}
}
\value{
The calculated APE metric.
}
\description{
This function calculates the accumulated prediction error (APE) from a fitted MXFAR model
}
\examples{
# Simulating from a bivariate MXFAR model with exponential coefficients
Fmat = function(x,r){
  f11 <- -0.3 + r[1]
  f12 <- 0.6*exp(-(0.30 + r[2])*x^2)
  f21 <- -0.2 + r[3]
  f22 <- -0.4*exp(-(0.45 + r[4])*x^2)
  return(list(matrix(c(f11,f21,f12,f22),ncol=2)))
}
data <- MXFAR.sim(SeriesNum = 10,Tlength = 500,d = 2,Y_d = 0,Fmat,
reff.Sigma = rep(0.05,4),eps.Sigma = rep(1,2),nburn = 500)

# Calculating APE for different model specifications
APE(SeriesNum = 10,Tlength = 500,y = data$y,u = data$ref,
p = 1,d = 2,r = 50,Q = 4) # best model
APE(SeriesNum = 10,Tlength = 500,y = data$y,u = data$ref,
p = 1,d = 3,r = 50,Q = 4)
APE(SeriesNum = 10,Tlength = 500,y = data$y,u = data$ref,
p = 2,d = 3,r = 50,Q = 4)
APE(SeriesNum = 10,Tlength = 500,y = data$y,u = data$ref,
p = 2,d = 4,r = 50,Q = 4)
}
