% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{FAR.est}
\alias{FAR.est}
\title{Multivariate FAR estimation}
\usage{
FAR.est(y, u, p, d, bwp = 0.1, numpoints = 50, fpdc = FALSE)
}
\arguments{
\item{y}{A matrix containing the multivariate time series data}

\item{u}{A vector containing the reference signal series}

\item{p}{The order of FAR(p) model}

\item{d}{The lag of the reference signal}

\item{bwp}{The bandwidth proportion from 0 to 1 (with respect to the range of observations)}

\item{numpoints}{The number of discretized points in the range of the reference signal 'u' that will be used to evaluate the local linear approximation}

\item{fpdc}{A boolean value to specify calculating the fPDC metric.}
}
\value{
A list of outputs including the functional coefficient estimates, the evaluation points and model residuals.
}
\description{
This function estimates the functional coefficients of a multivariate FAR model at specific reference signal value from a grid of evaluation points.
}
\examples{
# Simulating from a bivariate EXPAR model
Fmat = function(x,r){
  f11 <- -0.3 + r[1]
  f12 <- 0.6*exp(-(0.30 + r[2])*x^2)
  f21 <- -0.2 + r[3]
  f22 <- -0.4*exp(-(0.45 + r[4])*x^2)
  return(list(matrix(c(f11,f21,f12,f22),ncol=2)))
}
data <- FAR.sim(Tlength = 500,d = 2,Y_d = 0,Fmat,reff = rep(0,4),
eps.Sigma = rep(1,2),nburn = 500)

# Estimating the model
result <- FAR.est(y = data$y,u = data$ref,p = 1,d = 2)

# Plotting functional coefficient estimates
par(mfrow = c(2,2))
for(i in 1:2){
  for(j in 1:2){
    plot(result$fhat_points,result$fhat[i,j,],type = "l",
    ylab = paste("fhat",i,",",j,sep = ""))
  }
}
}
