% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FARest}
\alias{FARest}
\title{Multivariate FAR estimation (C++)}
\usage{
FARest(y, u, u0, p, d, bwp)
}
\arguments{
\item{y}{A matrix containing the multivariate time series data}

\item{u}{A vector containing the reference signal series}

\item{u0}{Specific value of reference signal for local approximation}

\item{p}{The order of FAR(p) model}

\item{d}{The lag of the reference signal}

\item{bwp}{The bandwidth proportion from 0 to 1 (with respect to the range of observations)}
}
\value{
A matrix of slope and intercept estimates (locally at u0).
}
\description{
This function estimates the functional coefficients of a multivariate FAR model at specific reference signal value u0.
}
\keyword{internal}
