% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{fPDC}
\alias{fPDC}
\title{Functional Partial Directed Coherence Measure}
\usage{
fPDC(Fmat, freq)
}
\arguments{
\item{Fmat}{An array of functional coefficient matrices evaluated at different values of the reference signal.}

\item{freq}{A numeric vector of frequencies in (0,0.5) (with length typically equal to half of the time series length)}
}
\value{
An array of PDC values evaluated for each entry of 'freq' and each entry of 'fpoints'
}
\description{
This function calculates the functional partial directed coherence (PDC) measure from the functional coefficient matrices of a fitted functional-coefficient autoregressive (FAR) model.
}
\examples{
# Simulating from a bivariate EXPAR model
Fmat = function(x,r){
  f11 <- -0.3 + r[1]
  f12 <- 0.6*exp(-(0.30 + r[2])*x^2)
  f21 <- -0.2 + r[3]
  f22 <- -0.4*exp(-(0.45 + r[4])*x^2)
  return(list(matrix(c(f11,f21,f12,f22),ncol=2)))
}
data <- FAR.sim(Tlength = 500,d = 2,Y_d = 0,Fmat,reff = rep(0,4),eps.Sigma = rep(1,2),nburn = 500)

# Estimating the model
result <- FAR.est(y = data$y,u = data$ref,p = 1,d = 2)

# Calculating fPDC
freq <- seq(0,0.5,length = 50)
fpdc <- fPDC(result$fhat,freq)

# Plotting fPDC for different levels of the reference signal
par(mfrow = c(2,2))
for(i in 1:2){
  for(j in 1:2){
    plot(freq,fpdc[i,j,,1],type = "l",ylim = c(0,1),
    ylab = paste("fPDC",i,",",j,sep = ""))
  }
}

for(i in 1:2){
  for(j in 1:2){
    plot(freq,fpdc[i,j,,51],type = "l",ylim = c(0,1),
    ylab = paste("fPDC",i,",",j,sep = ""))
  }
}
}
