% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{MXFAR.est}
\alias{MXFAR.est}
\title{Multivariate MXFAR estimation}
\usage{
MXFAR.est(
  SeriesNum,
  Tlength,
  y,
  u,
  p,
  d,
  bwp = 0.1,
  numpoints = 50,
  fpdc = FALSE
)
}
\arguments{
\item{SeriesNum}{A numeric vector stating the number of series in a 'group' data}

\item{Tlength}{A numeric scalar stating the number of time points per series}

\item{y}{A matrix containing the 'stacked' multivariate time series data from multiple series}

\item{u}{A vector containing the 'stacked' reference signal series}

\item{p}{The order of MXFAR(p) model}

\item{d}{The lag of the reference signal}

\item{bwp}{The bandwidth proportion from 0 to 1 (with respect to the range of observations)}

\item{numpoints}{The number of discretized points in the range of the reference signal 'u' that will be used to evaluate the local linear approximation}

\item{fpdc}{A boolean value to specify calculating the fPDC metric.}
}
\value{
A list of outputs including the functional coefficient (mean and series-specific) estimates, the evaluation points and model residuals.
}
\description{
This function estimates the functional coefficients of a multivariate mixed-effects FAR model at specific reference signal value from a grid of evaluation points.
}
\examples{
# Simulating from a bivariate MXFAR model with exponential coefficients
Fmat = function(x,r){
  f11 <- -0.3 + r[1]
  f12 <- 0.6*exp(-(0.30 + r[2])*x^2)
  f21 <- -0.2 + r[3]
  f22 <- -0.4*exp(-(0.45 + r[4])*x^2)
  return(list(matrix(c(f11,f21,f12,f22),ncol=2)))
}
data <- MXFAR.sim(SeriesNum = 10,Tlength = 500,d = 2,Y_d = 0,Fmat,
reff.Sigma = rep(0.05,4),eps.Sigma = rep(1,2),nburn = 500)

# Estimating the model
result <- MXFAR.est(SeriesNum = 10,Tlength = 500,
y = data$y,u = data$ref,p = 1,d = 2)

# Plotting mean functional coefficient estimates
par(mfrow = c(2,2))
for(i in 1:2){
  for(j in 1:2){
    plot(result$fhat_points,result$fhat_mean[i,j,,1],type = "l",
    ylab = paste("fhat",i,",",j,sep = ""))
  }
}
}
