% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MXFARest}
\alias{MXFARest}
\title{Multivariate MXFAR estimation (C++)}
\usage{
MXFARest(SeriesNum, Tlength, y, u, u0, p, d, bwp)
}
\arguments{
\item{SeriesNum}{A numeric vector stating the number of series in a 'group' data}

\item{Tlength}{A numeric scalar stating the number of time points per series}

\item{y}{A matrix containing the 'stacked' multivariate time series data from multiple series}

\item{u}{A vector containing the 'stacked' reference signal series}

\item{u0}{Specific value of reference signal for local approximation}

\item{p}{The order of MXFAR(p) model}

\item{d}{The lag of the reference signal}

\item{bwp}{The bandwidth proportion from 0 to 1 (with respect to the range of observations)}
}
\value{
A matrix of slope and intercept estimates (locally at u0).
}
\description{
This function estimates the functional coefficients of a multivariate mixed-effects FAR model at specific reference signal value u0.
}
\keyword{internal}
