% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{PDC}
\alias{PDC}
\title{Partial Directed Coherence Measure}
\usage{
PDC(phi, freq)
}
\arguments{
\item{phi}{A list of VAR coefficient matrices (with the model order as length of the list) or a 'column-wise stacked' matrix of VAR coefficients.}

\item{freq}{A numeric vector of frequencies in (0,0.5) (with length typically equal to half of the time series length)}
}
\value{
An array of PDC values evaluated for each entry of 'freq'
}
\description{
This function calculates the partial directed coherence (PDC) measure from the coefficient matrices of a fitted vector autoregressive (VAR) model.
}
\examples{
# Coefficients of a bivariate VAR model
phi1 <- matrix(c(0.4,-0.2,0,0.5),ncol = 2)
phi2 <- matrix(c(0.1,0.6,0,-0.25),ncol = 2)
phi <- list(phi1,phi2)

# Calculating PDC
freq <- seq(0,0.5,length = 50)
pdc <- PDC(phi,freq)

# Plotting PDC
par(mfrow = c(2,2))
for(i in 1:2){
  for(j in 1:2){
    plot(freq,pdc[i,j,],type = "l",ylim = c(0,1),ylab = paste("PDC",i,",",j,sep = ""))
  }
}
}
