% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{FAR.est}
\alias{FAR.est}
\title{Multivariate FAR estimation}
\usage{
FAR.est(y, u, p, d, bwp = 0.1, numpoints = 50)
}
\arguments{
\item{y}{A matrix containing the multivariate time series data}

\item{u}{A vector containing the reference signal series}

\item{p}{The order of FAR(p) model}

\item{d}{The lag of the reference signal}

\item{bwp}{The bandwidth proportion from 0 to 1 (with respect to the range of observations)}

\item{numpoints}{The number of discretized points in the range of the reference signal 'u' that will be used to evaluate the local linear approximation}
}
\value{
A list of outputs including the functional coefficient estimates, the evaluation points and model residuals.
}
\description{
This function estimates the functional coefficients of a multivariate FAR model at specific reference signal value u0.
}
